project/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
payu-websdk
pycryptodome==3.20.0
